Logging and post-maintenance updates
Logging will now include the last five lines of the log file of any
failed radmind processes for debugging. Logging also now notifies when
an image file is deleted.

Post-maintenance now overrides the HARD_DISK_NAME variable to ensure
that the specified name is used in place of anything else.